Move feedcachepolicy to its own file
