+++ EssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,22 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeedTests
+//
+//  Created by Gabriel Maldonado Almendra on 15/4/23.
+//
+
+import Foundation
+
+final class FeedCachePolicy {
+    private static let calendar = Calendar(identifier: .gregorian)
+    private static var maxCacheAgeInDays: Int { return 7 }
+    
+    private init() {}
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
